development update 5
-updated version number to v1.1.2-alpha
-added input parameter checking to the create new sample function to
make sure that the sample name is not blank and does not conflict with
any current samples in the database and the thickness is greater than 0.
-updated DoOpenMutliFileDialog function to not print out "cancelled" to
the command line when the file open operation is cancelled by the user
-updated the load TOF data files function to check the parameter input
to make sure that the specified voltage amplification is greater than 0,
and add checks for when the user cancels the parameter input
-corrected a bug in the load TOF data file function so that it does not
parse the header of the ToF data files using a comma delimiter
-updated the plot ToF Data function to extend the y-axis range of the
current transient to 0.01 down from 0.02 so taht the user can see more
of the data set.